requirements.txt
numpy
unyt
scipy
matplotlib
joblib
attrs
pytest
black
tqdm
velociraptor
swiftsimio
swiftgalaxy
pandas

sphinx
sphinx_rtd_theme
recommonmark

setuptools
h5py
astropy
hypothesis
PyYAML